feat: add waste item and waste type routes and services
- Created routes for managing waste items and waste types, including CRUD operations.
- Implemented services for waste item and waste type management, including validation and business logic.
- Added functionality to handle user roles and permissions for waste item and waste type operations.
- Introduced image upload handling for waste items.
- Updated server error handling to use consistent error handler.
- Implemented chat with AI service for user interactions.
- Added dropoff and transaction services for managing user dropoffs and transactions.
- Enhanced user service for user registration, login, and profile management.
