+++ src/routes/dropoff.routes.js
@@ -7,8 +7,8 @@ import {
   updateDropoffStatusController,
   cancelDropoffController,
   deleteDropoffController,
-} from "../controllers/dropoffController.js";
-import roleMiddleware from "../middlewares/auth.js";
+} from "../controllers/dropoff.controller.js";
+import roleMiddleware from "../middlewares/auth.middleware.js";
 
 const router = express.Router();
 
@@ -16,7 +16,11 @@ const router = express.Router();
 router.get("/", roleMiddleware("USER", "ADMIN"), getDropoffsController);
 
 // Get dropoffs for a specific user
-router.get("/users/:userId", roleMiddleware("USER", "ADMIN"), getUserDropoffsController);
+router.get(
+  "/users/:userId",
+  roleMiddleware("USER", "ADMIN"),
+  getUserDropoffsController
+);
 
 // Get a specific dropoff
 router.get("/:id", roleMiddleware("USER", "ADMIN"), getDropoffByIdController);
@@ -25,10 +29,18 @@ router.get("/:id", roleMiddleware("USER", "ADMIN"), getDropoffByIdController);
 router.post("/", roleMiddleware("USER", "ADMIN"), createDropoffController);
 
 // Update dropoff status (admin only)
-router.patch("/:id/status", roleMiddleware("ADMIN"), updateDropoffStatusController);
+router.patch(
+  "/:id/status",
+  roleMiddleware("ADMIN"),
+  updateDropoffStatusController
+);
 
 // Cancel a dropoff
-router.patch("/:id/cancel", roleMiddleware("USER", "ADMIN"), cancelDropoffController);
+router.patch(
+  "/:id/cancel",
+  roleMiddleware("USER", "ADMIN"),
+  cancelDropoffController
+);
 
 // Delete a dropoff (admin only)
 router.delete("/:id", roleMiddleware("ADMIN"), deleteDropoffController);
