+++ src/routes/index.js
@@ -1,9 +1,9 @@
-import userRoutes from "./userRoutes.js";
-import wasteTypeRoutes from "./wasteTypeRoutes.js";
-import dropoffRoutes from "./dropoffRoutes.js";
-import wasteItemRoutes from "./wasteItemRoutes.js";
-import transactionRoutes from "./transactionRoutes.js";
-import chatWithAitRoutes from "./chatWithAiRoutes.js";
+import userRoutes from "./user.routes.js";
+import wasteTypeRoutes from "./waste-type.routes.js";
+import dropoffRoutes from "./dropoff.routes.js";
+import wasteItemRoutes from "./waste-item.routes.js";
+import transactionRoutes from "./transaction.routes.js";
+import chatWithAitRoutes from "./chat-with-ai.routes.js";
 import { Router } from "express";
 
 const router = Router();
